requirements.txt
et-xmlfile==1.1.0
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
pillow==10.4.0
playsound==1.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==69.5.1
six==1.16.0
tk==0.1.0
tzdata==2024.1
wheel==0.43.0
